Use a base class to standardize functions.
